Fix remaining missing debugger build flags for open source builds (#54774)
Summary:

NOTE: This diff is a backport of https://github.com/facebook/react-native/pull/54770, where on the `0.83-stable` release branch, Network support for React Native DevTools was in a broken state under the open source build systems.

### Cause

Network debugging support depends a number of `REACT_NATIVE_DEBUGGER_ENABLED` preprocessor flags, which we use to compile away any overhead in production builds.

As we unfortunately use a total of **4 native build systems** today (with Buck 2 internally and primarily), the registration of these flags was missing across a number of native ObjC/C++ packages, which are now fixed with this PR.

- D87864636 aimed to address this as we weren't seeing the Network panel at all. However, it was insufficient, as it has only partially enabled network features between platforms.

### This diff

Add missing preprocessor flags in:

- Android:
	- `src/main/jni/react/devsupport/CMakeLists.txt`
	- `src/main/jni/CMakeLists.txt`
- iOS (Pods):
	- `React-jsinspectorNetwork`
- iOS (`Package.swift`):
	- `Libraries/Network`

Changelog: [Internal]

Reviewed By: vzaidman

Differential Revision: D88284345