Remove Commit from FrameTimingSequence
Summary:
# Changelog: [Internal]

This is not required for displaying the Frames track.

In Chrome, I've noticed that "Commit" is actually a complete event (`ph =  X`), and it has a `disabled-by-default-devtools.timeline`. Frames events, like `BeginFrame`, `DrawFrame`, `DroppedFrame` have `.frame` suffix in a category.

This event being a complete event actually makes lot of sense, since this is a work with some duration, where Host spent time committing the changes. In React Native case, this should probably reflect the `completeRoot()` call or some proxy from it. Basically the duration of committing tree operations in Fabric.

Differential Revision: D88382285