Introduce `AnimationBackendCommitHook` (#54138)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54138

This diff introduces `AnimationBackendCommitHook`. It is responsible for keeping animation updates in sync with React. When React does its updates, it doesn't look at the last mounted tree. Instead it hold its own references to ShadowNodes, and commits them whenever there is a (JS) rendering update. Any props that were commited from outside of React (e.g. by Reanimated) are lost.

To work around this we have the Commit Hook. It enables us to override any rendering updates that react does, just before the ShadowTree enters the layout phase.

In this implementation we utilize RNSRU to push those updates back to React (whenever there is a commit originating from React). So after a render React will now hold references to ShadowNodes with the current animation state (in props). This is a new approach, in Reanimated's current implementation this is disabled, and we reapply the changes via the Commit Hook on each React commit.

# Changelog
[General] [Added] - Add `AnimationBackendCommitHook` and `AnimatedPropsRegistry
[General] [Changed] - Use the commit hook to reconcile the UI updates with react

Reviewed By: sammy-SC, zeyap

Differential Revision: D84250600

fbshipit-source-id: c5e4b2030e55830ac5df29bde492d33a10f2de35