Reduce the screenshots size (#54800)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54800

# Changelog: [Internal]

From my observations, currently every base64 string for encoded screenshot takes on average ~100Kb. This is too much. These changes reduce the size to ~6Kb, at a cost of quality.

We can't accomodate such big strings, because there could be hundreds, if not thousands of frames for a ~10-20s trace.

Reviewed By: huntie

Differential Revision: D88489936

fbshipit-source-id: c9df916d80ea013391b1f353c9815196a6cde8dd