fix: RNGP using node invocation non-compatible with Gradle Compilation Cache (#54801)
Summary:
Invoking `Runtime.getRuntime().exec()` is not compatible with Gradle Configuration Cache and providers should be used instead.

This error hasn't surfaced yet due to fact that this branch of code is rarely hit as the users usually provide correct path to the CLI in their `app/build.gradle` file. I stumbled upon it accidentally when bumping RN in monorepo and the CLI path was no longer valid.

## Changelog:

[ANDROID] [FIXED] - RNGP using node invocation non-compatible with Gradle Compilation Cache

Pull Request resolved: https://github.com/facebook/react-native/pull/54801

Test Plan: I tested it in a rnc-cli app, where I provided an invalid path for `cliFile` - the error was fixed after applying this patch.

Reviewed By: mdvacca, huntie

Differential Revision: D88492310

Pulled By: cortinico

fbshipit-source-id: a91e47b90937945066bde938bdf16b2550322afb