Hide Performance Monitor when tracing in RNDT
Summary:
# Changelog: [Internal]

As titled. Before this change, monitor was just disabled, but was shown in the UI.

I think we should completely hide it, so it doesn't contribute to any frames or screenshots data. Since the trace is already initialized from the RNDT, user has already expressed the intent for tracing, we don't need to show it in the UI.

Differential Revision: D88748037