Implement onStateChange in TracingStateListener jni interface (#54824)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54824

# Changelog: [Internal]

The previous approach was causing a crash if we would attempt to register multiple listeners on Android side, because we would static-initialize different methods from different instances.

When the native code tried to invoke onStateChanged on a different listener class, JNI threw an error.

Instead, we will cache the interface method.

Reviewed By: javache

Differential Revision: D88734482

fbshipit-source-id: 27a0fe69979810c3ae04477a2529a581ec89e15e