Hide Performance Monitor when tracing in RNDT (#54828)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54828

# Changelog: [Internal]

As titled. Before this change, monitor was just disabled, but was shown in the UI.

I think we should completely hide it, so it doesn't contribute to any frames or screenshots data. Since the trace is already initialized from the RNDT, user has already expressed the intent for tracing, we don't need to show it in the UI.

Reviewed By: huntie

Differential Revision: D88748037

fbshipit-source-id: 899d25f25ac8c4633448d2132f99b478f8375265