codereview: refactor header file generator even more
Now it reads from podspec files, except for some special cases that we have in the config file.

Updated RCTSwiftUIWrapper.podspec to use podspec_sources (which we use to detect source)

I tested this against the header files I got with the previous iteration, and also with the ones installed by Cocoapods.