Move state of PackagerConnectionSettings to companion object (#54852)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54852

While working on internal apps, I've noticed that there are scenarios where more than one `PackagerConnectionSettings` is instantiated.

What that happens, we're not able to reliability change the host anymore, because the state is scattered through the various `PackagerConnectionSettings`. Instead the state should be retained as a singleton.

Changelog:
[Internal] [Changed] - Move state of PackagerConnectionSettings to companion object

Reviewed By: vzaidman

Differential Revision: D88945232

fbshipit-source-id: b416e39e30b3b2b41cb0fcac60a700cc716d3c62