Implement Fabric commit branching (behind a flag) (#54835)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54835

Changelog: [General][Changed] - Changed how React changes are commited to the Shadow Tree (behind a feature flag)

Introduces two branches for the Shadow Tree commits - React always commits to the JS revision, while other sources should commit to the main revision. Commits to the JS revision schedule a "merge" commit at the end of the event loop pass to upstream the React changes to the main revision.

"Merge" means taking the JS revision and promoting it to become the new main, losing changes that happened between fork and merge. Preservation of these changes is the responsibility of the party that created them, and they should be reapplied using a commit hook.

Differential Revision: D88151489