Schedule JS revision merge on the main thread (#54837)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54837

Changelog: [General][Changed] - The JS revision merge now is scheduled on the main thread

Changes the Shadow Tree branching to schedule JS revision merge on the main thread instead of running it immediately at the end of the event loop. This will prevent shadow tree exhaustion when committing to the main branch, since all commits there will be happening on the same thread.

Differential Revision: D88736123