fixed issue with xcconfig being overwritten (#54838)
Summary:
To be able to change the target xcconfig from within `add_rn_core`, we need to make sure it is not overwritten by the line that sets the config.

This commit fixes this by moving it after the line:

`spec.pod_target_xcconfig = current_config`

## Changelog:

[IOS] [FIXED] - Moved setting xcconfig to after `add_rn_core` pod function is called.

Pull Request resolved: https://github.com/facebook/react-native/pull/54838

Test Plan: Run RNTester with precompiled binaries.

Reviewed By: huntie

Differential Revision: D88849730

Pulled By: cipolleschi

fbshipit-source-id: 632e23f5a464ea779035ca3e090c0a3c9475a260