Fix mask-based clipping for < API 27 (#54914)
Summary:

The previous implementation for path clipping on API 27 would only attempt to draw a clipping mask within the bounds of the inner content, applying anti-aliasing to the edge of that padding box path. However, it would neglect to account for outlines, resulting in drawing inner contents *outside* the boundary of the box. 

This change inverts the logic by applying a mask to the entire content boundary and using DST_OUT to "delete" content outside of the padding box, while also still honoring the padding box path and its anti-aliasing

Changelog: [Internal]

Differential Revision: D89324397