Gate requestBody processing behind BuildConfig.DEBUG (#54920)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54920

Gates `requestBody` handling for `reportRequestStart` on Android with `BuildConfig.DEBUG`. This is an appropriate optimisation as we do not use body previews outside of CDP debugging, aligning with the same conditional behaviour on iOS.

I'd like to also pick this change for 0.83.1, with D89373824, for additional safety.

**Other notes**

Gating by `BuildConfig` isn't done elsewhere in this file, as other potentially expensive behaviour is gated lower down by both `REACT_NATIVE_DEBUGGER_ENABLED` and the `Network.enable` CDP message. This was the only remaining gap with the iOS implementation.

Changelog: [Internal]

Reviewed By: vzaidman

Differential Revision: D89377163

fbshipit-source-id: 10a27f6b02f65755d4293f5cf33697d542bc9465