Add SurfaceHandlerCommitHook responsible for preserving layout constraints (#54836)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54836

Changelog: [General][Added] Added a `SurfaceHandlerCommitHook` to preserve layout constraints during commits from different threads

Since layout constraints and layout context are being kept in the props of the `RootShadowNode`, it was possible for the JS revision merge to lose an update if it was being commited concurrently.

`SurfaceHandlerCommitHook` fixes this by ensuring that the root node always uses the latest context and constraints.

Differential Revision: D88839901