fix(android): handle transitioning views (#130)
* first impl

* add UIManager methods

* add custom view manager to test code

* paper implement method

* refactor code: make ViewTransitioNCoordinator handle the main stuff

* fix tests with discord changes

* test for viewtransitioncoordinator

* add test to verify ViewTransitionCoordinator

* fix various edge cases

* test case

* remove debug logs

* wip: add deleteView to queue as well

* refactor: make childTag first arg

* removed checked todo item

* fix queue getting stuck

* fix correctness of queues clearing up properly

* avoid confusing double loop

* cleanup previous mechanisms

* refactor: exec. first removes immediately

There were two mistake in the previous implementation.
Let's say we have:
- markTransition A
- remove A
- delete A

then we need to execute the remove A immediately, instead of queuing it. A will be removed from its parent, but as its in transition, will keep its pointer to the parent.
This will cause the onDetach listener to fire naturally.

The second mistake was to drain the queue when markViewInTransition=false was being called.
This will be called e.g. during drawing passes (ie. on animation end), and we don't want to change the view hierarchy during draw passes!
With the previous fix the onDetach should always be called, so running it in markViewInTransition is not necessary any longer

* add explicit handling for when view has parent but markViewInTransition was not called