Move ShadowNodeFamily to PropsAnimatedNode (#54879)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54879

This diff is a part of the process of getting the Animated-itest to work with Animation Backend. During testing I found that sometimes the disconnect method would cleanup `tagToShadowNodeFamily_` when there were more than one PropsAnimatedNode for a view, so we would wrongly skip an animation. By storing the family pointer on the props node we can avoid that.

Differential Revision: D89042963