Updating frontend assets to add header for open-in-editor sources panel devtools button (#54927)
Summary:
Updating front end assets for https://github.com/facebook/react-native-devtools-frontend/pull/229

The "Open in Editor" button in the Sources tab was failing with a 400 Bad Request error. This was because the openStackFrameInEditor request was missing the Content-Type header. Added the appropriate content type header to fix the request.

## Changelog:

[GENERAL] [INTERNAL] - Fix "Open in Editor" button in Sources tab returning 400 Bad Request

Pull Request resolved: https://github.com/facebook/react-native/pull/54927

Test Plan:
Open React Native DevTools
Go to Sources tab
Click "Open in Editor" button on a stack frame
Verify the file opens in the editor without a 400 error

Reviewed By: vzaidman

Differential Revision: D89464184

Pulled By: emily8rown

fbshipit-source-id: 27190e05e11d96bdd9982a4e057c3ef225975bc4