Resolve `use_react_native!(path)` param from pwd
The `path` argument in `react_native_pods.rb` is not resolved but
instead later joined using `File.join(relative_path_from_current,
path)`. This doesn't actually resolve the path, meaning, if `path` is
absolute it's appended to this first relative path.

Instead, we should use `Pathname#join` and join the path to the
pathname, guaranteeing it to be absolute, then construct relative paths
(for the prefix) using `relative_path_from` against the installation
root.