Add Flow lib defs for newer Node.js fs APIs
Summary:
Node.js's `fs` has been updated considerably since it was first covered with Flow, and so far we've only added new APIs piecemeal as we need them.

This goes through the whole module top to bottom and attempts to update in line with the Node.js docs at v24.

Changelog: [Internal]

Reviewed By: vzaidman

Differential Revision: D89605689