Add Flow lib defs for newer Node.js fs APIs (#54992)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54992

Node.js's `fs` has been updated considerably since it was first covered with Flow, and so far we've only added new APIs piecemeal as we need them.

This goes through the whole module top to bottom and attempts to update in line with the Node.js docs at v24.

Changelog: [Internal]

Reviewed By: vzaidman

Differential Revision: D89605689

fbshipit-source-id: 1ad0ef2f7fedb6416471461b8be7bc418afcd28f